IMS/Inventory_management_system/requirements.txt
Django==5.1.5
django-auto-prefetch==1.14.0
dj-database-url==1.3.0
python-dotenv==1.2.1
openpyxl==3.1.5
Pillow==12.1.0
psycopg2-binary==2.9.11
django-crispy-forms==2.5
crispy-bootstrap5==2025.6
matplotlib==3.10.8
pandas==3.0.0
reportlab==4.4.9
seaborn==0.13.2
xlsxwriter
qrcode[pil]==8.2
gunicorn==25.0.3
whitenoise==6.11.0
plotly==6.5.2

# Microsoft 365 Integration
msal==1.31.0
requests>=2.32.3
cryptography>=44.0.0
